Update departments.ps1
In stead of calling a piece of script before a while loop, and then repeating the same script in the while loop, you can use a do while loop.
This way you prevent redundancy and needing to change a script in 2 places but the effect is the same.

We checked if this works (Geo and me). Please review if this is best practice.